third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/pseudo-elements-008-expected.txt

This is a testharness.js-based test.
PASS Originating element container for ::before
PASS Originating element container for ::after
PASS Originating element container for ::marker
PASS Originating element container for ::first-line
PASS Originating element container for ::first-letter
PASS Originating element container for outer ::first-line
PASS Originating element container for outer ::first-letter
FAIL Originating element container for ::backdrop assert_equals: ::backdrop not rendered expected "30px" but got "80px"
Harness: the test ran to completion.
